requirements.txt
gitdb2==2.0.0
PyYAML==3.12
requests==2.12.4
sh==1.12.9
smmap2==2.0.1
